requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
idna==3.2
psycopg2==2.9.1
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
spotipy==2.19.0
sqlparse==0.4.1
tornado==6.1
tzlocal==3.0
urllib3==1.26.6
xmltodict==0.12.0
